refactor(shared): declare as open module
The shared utility classes might be separated in subpackages for better
structuring. A closed module only exports the top-level classes, though.
We allow exposing the subpackages for this shared module to allow others
to access them.

If the shared module grows too big in the future, it might make more
sense to split it into dedicated modules that then are no longer marked
as open.